chore: improving types
